getNAImeta/cloud_function/requirements.txt
functions-framework==3.*
Pillow==10.*
numpy==1.* 